requirements.txt
# Core build tools
setuptools>=68
wheel

# Web app frameworks
streamlit==1.39.0
fastapi==0.115.0
uvicorn==0.30.6
requests>=2.31.0

# AI & ML
google-generativeai>=0.8.3
chromadb==0.4.24
vaderSentiment==3.3.2
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3

# LangChain core (optional lightweight)
langchain-core>=0.1.53
langchain-community>=0.0.38

# Environment variables
python-dotenv==1.0.1
